+++ mytimer/params.py
@@ -187,6 +187,18 @@ PROGRAMS_MAP = {
         "second": 0,
         "message": "Purple tea (3 mins, 80-85 C)",
     },
+    "mate": {
+        "hour": 0,
+        "minute": 5,
+        "second": 0,
+        "message": "Mate (5 mins, 65-70 C)",
+    },
+    "french-press": {
+        "hour": 0,
+        "minute": 4,
+        "second": 0,
+        "message": "French press (4 mins, 93-96 C)",
+    },
     "work": {
         "hour": 0,
         "minute": 25,